jobportal/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-allauth==0.58.2
scikit-learn==1.3.2
nltk==3.8.1
python-dotenv==1.0.0
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.0.1
gunicorn==21.2.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
openai==1.3.0
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary
